==> Python_data_analysis_project/content_engagement_churn_analytics/notebooks/01_data_generation_cleaning.ipynb <==
import pandas as pd
import numpy as np
from datetime import timedelta


np.random.seed(42)


# Users & content
n_users = 1200
content_types = ["Movie", "Series", "Sports", "Kids", "News"]
plans = ["Free", "Basic", "Premium"]

# Time range
start_date = "2023-01-01"
end_date = "2024-12-31"
dates = pd.date_range(start=start_date, end=end_date, freq="D")


users = []

for i in range(1, n_users + 1):
    signup_date = np.random.choice(dates[:500])
    plan = np.random.choice(plans, p=[0.45, 0.35, 0.20])

    users.append([
        f"U{i:05d}",
        signup_date,
        plan
    ])

user_df = pd.DataFrame(
    users,
    columns=["user_id", "signup_date", "subscription_plan"]
)


user_df.head()

data = []

for _, row in user_df.iterrows():
    user_id = row["user_id"]
    plan = row["subscription_plan"]
    signup_date = row["signup_date"]

    active_days = np.random.randint(90, 600)
    user_dates = pd.date_range(
        start=signup_date,
        periods=min(active_days, len(dates)),
        freq="D"
    )

    for date in user_dates:
        sessions = np.random.randint(1, 4)

        for _ in range(sessions):
            content_type = np.random.choice(content_types)
            watch_minutes = np.random.randint(5, 120)

            # Revenue logic
            if plan == "Free":
                ad_views = np.random.randint(1, 6)
                revenue = ad_views * 0.03
            elif plan == "Basic":
                ad_views = 0
                revenue = 0.25
            else:  # Premium
                ad_views = 0
                revenue = 0.40

            data.append([
                user_id,
                date,
                f"C{np.random.randint(1000, 9999)}",
                content_type,
                watch_minutes,
                sessions,
                plan,
                ad_views,
                revenue
            ])


engagement_df = pd.DataFrame(
    data,
    columns=[
        "user_id",
        "session_date",
        "content_id",
        "content_type",
        "watch_minutes",
        "session_count",
        "subscription_plan",
        "ad_views",
        "revenue"
    ]
)


engagement_df.head()

last_activity = (
    engagement_df
    .groupby("user_id")["session_date"]
    .max()
    .reset_index()
    .rename(columns={"session_date": "last_active_date"})
)

engagement_df = engagement_df.merge(
    last_activity,
    on="user_id",
    how="left"
)

cutoff_date = engagement_df["session_date"].max() - timedelta(days=30)

engagement_df["churn_flag"] = np.where(
    engagement_df["last_active_date"] < cutoff_date,
    1,
    0
)


engagement_df.head()


engagement_df.info()


engagement_df["churn_flag"].value_counts()


engagement_df.to_csv(
    "C:/Users/Abhi/Desktop/Python_Data_Analytics_Projects/content_engagement_churn_analytics/data/raw/content_engagement_raw.csv",
    index=False
)
