==> Python_data_analysis_project/content_engagement_churn_analytics/notebooks/02_engagement_feature_engineering.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv( "C:/Users/Abhi/Desktop/Python_Data_Analytics_Projects/content_engagement_churn_analytics/data/raw/content_engagement_raw.csv",parse_dates=["session_date", "last_active_date"])

df.head()

df = df.sort_values(["user_id", "session_date"])
df.head()

daily_df = (
    df.groupby(["user_id", "session_date", "subscription_plan", "churn_flag"])
      .agg(
          total_watch_minutes=("watch_minutes", "sum"),
          total_sessions=("session_count", "sum"),
          total_ads=("ad_views", "sum"),
          daily_revenue=("revenue", "sum")
      )
      .reset_index()
)


daily_df.head()

daily_df["watch_7d_avg"] = (
    daily_df.groupby("user_id")["total_watch_minutes"]
            .transform(lambda x: x.rolling(7, min_periods=1).mean())
)
daily_df.head()

daily_df["watch_30d_avg"] = (
    daily_df.groupby("user_id")["total_watch_minutes"]
            .transform(lambda x: x.rolling(30, min_periods=1).mean())
)
daily_df.head(15)

daily_df["engagement_volatility"] = (
    daily_df.groupby("user_id")["total_watch_minutes"]
            .transform(lambda x: x.rolling(14, min_periods=3).std())
)
daily_df.head(15)

daily_df["sessions_per_day"] = daily_df["total_sessions"]
daily_df.head()

daily_df["revenue_30d"] = (
    daily_df.groupby("user_id")["daily_revenue"]
            .transform(lambda x: x.rolling(30, min_periods=1).sum())
)
daily_df.head()

daily_df["engagement_change"] = (
    daily_df.groupby("user_id")["watch_7d_avg"]
            .diff()
)
daily_df.head()

user_features = (
    daily_df.groupby("user_id")
    .agg(
        subscription_plan=("subscription_plan", "last"),
        churn_flag=("churn_flag", "max"),
        avg_watch_minutes=("total_watch_minutes", "mean"),
        avg_sessions=("sessions_per_day", "mean"),
        engagement_volatility=("engagement_volatility", "mean"),
        avg_watch_7d=("watch_7d_avg", "mean"),
        avg_watch_30d=("watch_30d_avg", "mean"),
        revenue_30d=("revenue_30d", "mean"),
        engagement_trend=("engagement_change", "mean")
    )
    .reset_index()
)


user_features.head()

user_features["engagement_score"] = (
    0.4 * user_features["avg_watch_minutes"] +
    0.3 * user_features["avg_sessions"] -
    0.3 * user_features["engagement_volatility"]
)


user_features.head()


user_features.describe()


user_features["churn_flag"].value_counts()


user_features.to_csv(
     "C:/Users/Abhi/Desktop/Python_Data_Analytics_Projects/content_engagement_churn_analytics/data/processed/engagement_features.csv",
    index=False
)
